% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spatial_auc.R
\name{calculate_spatial_auc}
\alias{calculate_spatial_auc}
\title{Spatial prioritization of the response to a biological perturbation}
\usage{
calculate_spatial_auc(
  input,
  meta,
  label_col = "label",
  barcode_col = "barcode",
  coord_cols = c("x", "y"),
  nn = NULL,
  distance_metrics_df = NULL,
  dismay_methods = c("pearson", "spearman", "binomial"),
  k = 50,
  n_subsamples = 10,
  max_barcodes = 1000,
  barcodes_per_step = 10,
  epsilon = 5e-04,
  steps_tracking = 5,
  cor_method = "pearson",
  min_cor = 0.8,
  seed = 42,
  ncores = 8,
  save_feature_importance = F,
  test_sim = F,
  ...
)
}
\arguments{
\item{input}{a matrix containing gene expression values
(genes in rows, cells in columns); columns matching the barcodes in meta}

\item{meta}{a data frame containing metadata about the \code{input}
gene-by-cell matrix, containing the minimum of the coordinates columns specified (default \code{c('x', 'y')}), \code{barcode_col} column and \code{label_col} column}

\item{label_col}{the column of the \code{meta} data frame that
contains condition labels (e.g., disease, timepoint); defaults to \code{label}}

\item{barcode_col}{the column of the \code{meta} data frame that
contains barcode; defaults to \code{barcode}}

\item{coord_cols}{the coordinate columns that are found in metadata; defaults to
\code{('x', 'y')}}

\item{nn}{the nearest neighbors data frame derived from \link{get_nn}}

\item{distance_metrics_df}{distance metrics data frame derived from \link{get_distance_metrics}}

\item{dismay_methods}{distance metrics tures to calculate from \href{https://github.com/skinnider/dismay}{dismay} package}

\item{k}{(1) the nearest neighbors to calculate distance metrics for and (2) number of neighbours to use for Augur; defaults to \code{50}.}

\item{n_subsamples}{the number of random subsamples of fixed size to draw from the data; defaults to \code{10}.}

\item{max_barcodes}{hyperparamter for convergence; the maximum number of barcodes to run Augur on before stopping; defaults to \code{1000}.}

\item{barcodes_per_step}{hyperparamter for convergence; the number of barcodes to run Augur on per convergence step, analogous to learning rate; defaults to \code{10}.}

\item{epsilon}{hyperparamter for convergence; \eqn{\epsilon}, maximum \eqn{\Delta} between correlation between convergence steps; defaults to \code{5e-4}.}

\item{steps_tracking}{hyperparamter for convergence; number of convergence steps to track to assume convergence --  converged if number of steps with \eqn{\Delta} < \eqn{\epsilon}; defaults to \code{5e-4}.}

\item{cor_method}{correlation method to use for convergence; defaults to \code{pearson}}

\item{min_cor}{minimum correlation before assumption of convergence; defaults to \code{0.8}}

\item{seed}{seed for pseudorandom sampling; defaults to \code{42}}

\item{ncores}{number of cores to use, mainly for \code{Augur}; defaults to \code{8}}

\item{save_feature_importance}{boolean to save feature importance per AUC calculated; defaults to false}

\item{test_sim}{set to true only if testing Vespucci; load pre-loaded calculated_auc_df for spatial_sim data; defaults to false}
}
\value{
a list containing the following elements:
\itemize{
  \item \code{calculated_auc_df}: returns a data frame containing \code{barcode} and corresponding \code{auc} from \code{Augur::calculate_auc} that are used as training data
  \item \code{distance_metrics_df}: if argument \code{distance_metrics_df} is \code{NULL}, returns a data frame from \link{get_distance_metrics}; else if \code{distance_metrics_df} is not \code{NULL}, returns the same data frame as given in the argument
  \item \code{aucs}: a data frame with a \code{barcode} column and \code{auc} column after convergence
  \item \code{auc_tracking}: a data frame that tracks the aucs over the convergence step; contains a \code{barcode} column and their corresponding \code{auc} columns calculated for each convergence step
  \item \code{cor_tracking}: a data frame that tracks the correlation over convergence steps; contains \code{step}, \code{cor_method}: correlation method used and \code{cor}: correlation between step \code{i} and step \code{i-1}

  \item \code{time_tracking}: a list that contains
    \itemize{
         \item \code{global}: time taken for \link{get_nn} and \link{get_distance_metrics}
         \item \code{aucs}: time taken for \link{calculate_auc} for each barcode
         \item \code{cor_convergence}: time taken for building model and predicting AUCs for each convergence step
     }
  \item \code{feature_importance}: only if save_feature_importance is TRUE -- refer to https://github.com/neurorestore/Augur/blob/master/R/calculate_auc.R output
}
}
\description{
Prioritize spatial barcodes involved in a complex biological process by training (1) a machine-learning model to predict sample labels (e.g., disease vs. control, treated vs. untreated, or time post-stimulus), and evaluate the performance of the model in cross-validation framework, and (2) a meta-learning model that trains on predictions from the prior machine learning models and distance metrics obtained from the spatial transcriptomics.
}
