% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countsplit_matrix}
\alias{countsplit_matrix}
\title{Count split matrix}
\usage{
countsplit_matrix(input, seed = 42)
}
\arguments{
\item{input}{a matrix containing gene expression values
(genes in rows, cells in columns)}

\item{seed}{seed for pseudorandom sampling; defaults to \code{42}}
}
\value{
a list containing two matrices (one for running \code{calculate_spatial_auc}, the other for running \code{find_spatial_de})
}
\description{
split count matrix with Poisson spliting
}
