% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_spatial_de.R
\name{find_spatial_de}
\alias{find_spatial_de}
\title{Spatially relevant differential gene/GO term analysis}
\usage{
find_spatial_de(
  input,
  meta,
  auc_df,
  de_method = "nbgmm",
  barcode_col = "barcode",
  replicate_col = "replicate",
  go_list = NULL,
  ...
)
}
\arguments{
\item{input}{a matrix containing gene expression values
(genes in rows, cells in columns); columns matching the barcodes in meta}

\item{meta}{a data frame containing metadata about the \code{input}
gene-by-cell matrix, containing the minimum of the \code{barcode_col} column and \code{replicate_col} column}

\item{auc_df}{a data frame containing \code{barcode} (matching meta \code{barcode_col} column) and \code{auc} (AUC values) columns}

\item{de_method}{method to run differential expression (defaults to \code{nbgmm}).
\describe{
  \item{\code{nbgmm}}{Negative binomial generalized mixed model}
  \item{\code{glm}}{Generalized linear model}
}}

\item{barcode_col}{the column of the \code{meta} data frame that
contains barcode; defaults to \code{barcode}}

\item{replicate_col}{the column of the \code{meta} data frame that
contains the biological replicate id; defaults to \code{replicate}}

\item{go_list}{A list of vectors of features for expression programs; each entry should be a vector of feature names; used as input into \code{Seurat::AddModuleScore}}
}
\value{
a data frame containing the following columns:
\itemize{
  \item \code{feature}: gene/GO term
  \item \code{effect_size}: effect size based on AUC values
  \item \code{p_val}: p-value
  \item \code{p_val_adj}: BH adjusted p-value
}
}
\description{
Find differentially expressed genes/GO terms that corresponds to AUC derived from \code{navigate_space} function. This is calculated by fitting either a negative binomial generalized mixed model or a generalized linear mixed model  of the expression values to the AUC values while accounting for biological replicates.
}
