% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_distance_metrics}
\alias{get_distance_metrics}
\title{Distance metrics within spatial transcriptomics}
\usage{
get_distance_metrics(
  input,
  meta,
  nn,
  k = 50,
  label_col = "label",
  barcode_col = "barcode",
  dismay_methods = c("pearson", "spearman", "binomial"),
  select_var = T,
  ncores = 8
)
}
\arguments{
\item{input}{a matrix containing gene expression values
(genes in rows, cells in columns); columns matching the barcodes in meta}

\item{meta}{a data frame containing metadata about the \code{input}
gene-by-cell matrix, containing the minimum of the coordinates columns specified
(default \code{c('x', 'y')}), \code{barcode_col} column and \code{label_col} column}

\item{nn}{the nearest neighbors data frame derived from \link{get_nn}}

\item{k}{the nearest neighbors to calculate distance metrics for; defaults to
\code{50}.}

\item{label_col}{the column of the \code{meta} data frame that
contains condition labels (e.g., disease, timepoint); defaults to \code{label}}

\item{barcode_col}{the column of the \code{meta} data frame that
contains barcode; defaults to \code{barcode}}

\item{dismay_methods}{distance metrics to calculate from \href{https://github.com/skinnider/dismay}{dismay} package}

\item{select_var}{boolean to run \code{Augur}::select_variance, defaults to \code{True}}

\item{ncores}{number of cores to use}
}
\value{
a data frame containing the following columns:
\itemize{
  \item \code{barcode}: the barcode of each cell
  \item \code{dismay summary statistics}: for each dismay method in \code{dismay_methods}, the following summary
                 statistics will be calculated
                 \itemize{
                     \item \code{mean}
                     \item \code{median}
                     \item \code{sd}: standard deviation
                     \item \code{q1}: 25th quantile
                     \item \code{q3}: 75th quantile
                 }
}
}
\description{
Get the distance metrics of each barcode to its surrounding nearest neighbors
}
