% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_GO_matrix}
\alias{get_GO_matrix}
\title{Get Gene Ontology (GO) level score matrix}
\usage{
get_GO_matrix(input, go_list, ncores = 8)
}
\arguments{
\item{input}{a matrix containing gene expression values
(genes in rows, cells in columns); columns matching the barcodes in meta}

\item{go_list}{A list of vectors of features for expression programs; each entry should be a vector of feature names; used as input into \code{Seurat::AddModuleScore}}

\item{ncores}{number of cores to use}
}
\value{
a matrix containing GO level scores
}
\description{
Get the GO score matrix from expression matrix
}
