% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nn}
\alias{get_nn}
\title{Derive nearest neighbors within spatial transcriptomics}
\usage{
get_nn(
  meta,
  nn_radius = 1000,
  coord_cols = c("x", "y"),
  label_col = "label",
  barcode_col = "barcode"
)
}
\arguments{
\item{meta}{a data frame containing metadata of the spatial transcriptomics object,
containing the minimum of the coordinates columns specified (default
\code{c('x', 'y')}), \code{barcode_col} column and \code{label_col} column}

\item{nn_radius}{determines the radius of the nearest neighbors to search for; defaults to \code{1000}.}

\item{coord_cols}{the coordinate columns that are found in metadata; defaults to
\code{('x', 'y')}}

\item{label_col}{the column of the \code{meta} data frame that
contains condition labels (e.g., disease, timepoint); defaults to \code{label}}

\item{barcode_col}{the column of the \code{meta} data frame that
contains barcode; defaults to \code{barcode}}
}
\value{
a data frame containing the following columns:
\itemize{
  \item \code{source_barcode}: the source barcode
  \item \code{source_label}: the label that the source barcode belongs to
  \item \code{target_barcode}: the target barcode
  \item \code{target_group}: the label that the target barcode belongs to
  \item \code{rank}: the rank based on the nearest neighbors within each label,
     the lower the number, the closer the target barcode is to the source barcode
  \item \code{neighbor_idx}: the rank based on the nearest neighbors,
     the lower the number, the closer the target barcode is to the source barcode
}
}
\description{
Find nearest neighbors of each barcode in a spatial transcriptomic object
}
