% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vespucci.R
\name{run_vespucci}
\alias{run_vespucci}
\title{Spatial prioritization of the response to a biological perturbation and identification of spatially relevant features}
\usage{
run_vespucci(input, meta, seed = 42, ...)
}
\arguments{
\item{input}{a matrix containing gene expression values
(genes in rows, cells in columns); columns matching the barcodes in meta}

\item{meta}{a data frame containing metadata about the \code{input}
gene-by-cell matrix, containing the minimum of the coordinates columns specified (default \code{c('x', 'y')}), \code{barcode_col} column, \code{replicate_col} column and \code{label_col} column}

\item{seed}{seed for pseudorandom sampling; defaults to \code{42}}
}
\value{
a list containing the following elements:
\itemize{
  \item \code{spatial_auc_result}: list output from \code{calculate_spatial_auc}; most relevant element is \code{spatial_auc_results$aucs}, which is a barcode-to-auc data frame.
  \item \code{de_feature_result}: data frame output from \code{find_spatial_de}
}
}
\description{
Vespucci first prioritize spatial barcodes involved in a complex biological process by training (1) a machine-learning model to predict sample labels (e.g., disease vs. control, treated vs. untreated, or time post-stimulus), and evaluate the performance of the model in cross-validation framework, and (2) a meta-learning model that trains on predictions from the prior machine learning models and distance metrics obtained from the spatial transcriptomics. Finally, Vespucci identifies spatially relevant features by fitting the expression and AUC values to a negative binomial generalized linear mixed model or a generalized linear mixed model. \cr Essentially \code{run_vespucci} runs the following functions in this particular order: \code{countsplit_matrix}, \code{calculate_spatial_auc} and \code{find_spatial_de}.
}
